Move django adapter to `adapters`
